+++ Cargo.toml
@@ -17,5 +17,13 @@ atomic = "0.4.5"
 spin = "0.5.2"
 thread_local = "1.0.1"
 
+
 [target.'cfg(windows)'.dependencies]
 winapi = { version = "0.3.8", features=["std", "winuser", "heapapi", "ntdef", "winnt", "winerror", "memoryapi"] }
+
+[dev-dependencies]
+criterion = "0.3"
+
+[[bench]]
+name = "allocating_bench"
+harness = false
