--- no_harness_test/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "no_harness_test"
-version = "0.1.0"
-authors = ["Joshua Radin <jradin16@gmail.com>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-lralloc-rs = {version = "*", path="../"}
